Better diagnostics in AbortProcess(Silently|Dramatically) commit_hash:7bbc13afbd569ca9bb064aed656c12644fd2de96

+++ yt/yt/server/job_proxy/job_proxy.cpp
@@ -1776,7 +1776,7 @@ void TJobProxy::Abort(EJobProxyExitCode exitCode)
         job->Cleanup();
     }
 
-    TProgram::Abort(static_cast<int>(exitCode));
+    TProgram::Abort(exitCode);
 }
 
 bool TJobProxy::TrySetCpuGuarantee(double cpuGuarantee)
